Bring up gem to a newer version of ruby

+++ lib/paperclip/storage/backblaze.rb
@@ -24,7 +24,7 @@ module Paperclip
     #   class Note < ApplicationRecord
     #     has_attached_file :image,
     #       storage: :backblaze,
-    #       b2_credentials: Rails.root.join('config/b2.yml'),
+    #       b2_credentials: YAML.safe_load(ERB.new(File.read("#{Rails.root}/config/back_blaze.yml")).result)[Rails.env],
     #     ...
     #
     # You can put the backblaze config in the environment config file e.g.:
@@ -33,44 +33,36 @@ module Paperclip
     #
     #   config.paperclip_defaults = {
     #     storage: :backblaze,
-    #     b2_credentials: Rails.root.join('config/b2.yml')
+    #     b2_credentials: [Hash]
     #   }
     #
     # If you do it this way, then you don't need to put it in the model.
     #
     module Backblaze
-      def self.extended base
+      def self.extended(base)
         base.instance_eval do
           login
-          unless @options[:url].match(/\Ab2.*url\z/)
-            @options[:url] = ":b2_path_url".freeze
+          unless @options[:url] =~ /\Ab2.*url\z/
+            @options[:url] = ':b2_path_url'.freeze
           end
         end
 
-        Paperclip.interpolates(:b2_path_url) do |attachment, style|
-          "#{::Backblaze::B2.download_url}/file/#{attachment.b2_bucket_name}/#{attachment.path(style).sub(%r{\A/}, '')}"
-        end unless Paperclip::Interpolations.respond_to? :b2_path_url
+        unless Paperclip::Interpolations.respond_to? :b2_path_url
+          Paperclip.interpolates(:b2_path_url) do |attachment, style|
+            "#{::Backblaze::B2.download_url}/file/#{attachment.b2_bucket_name}/#{attachment.path(style).sub(%r{\A/}, '')}"
+          end
+        end
       end
 
-      # Fetch the credentials from the config file, if it hasn't already been
-      # loaded.
-      #
-      # filename : String - path to the YAML config file containing the
-      # Backblaze B2 credentials. Here is example contents of what one
-      # may look like:
-      #
-      #   account_id: 123456789abc
-      #   application_key: 0123456789abcdef0123456789abcdef0123456789
-      #
+      # Reads b2_credentials from the config file
+      # must be a hash
+      #   {
+      #     account_id: 123456789abc
+      #     application_key: 0123456789abcdef0123456789abcdef0123456789
+      #   }
       # Returns a Hash containing the parsed credentials.
-      def b2_credentials(filename = @options[:b2_credentials])
-        unless @b2_credentials
-          require 'psych'
-          File.open(filename, 'r') do |f|
-            @b2_credentials = Psych.load(f.read).symbolize_keys
-          end
-        end
-        @b2_credentials
+      def b2_credentials
+        @b2_credentials ||= @options.fetch(:b2_credentials)
       end
 
       # Authenticate with Backblaze with the account ID and secret key. This
@@ -79,13 +71,18 @@ module Paperclip
       def login
         return if ::Backblaze::B2.token
         creds = b2_credentials
-        ::Backblaze::B2.login(account_id: creds[:account_id], application_key: creds[:application_key])
+        ::Backblaze::B2.login(
+          account_id: creds.fetch(:account_id),
+          application_key: creds.fetch(:application_key)
+        )
       end
 
       # Return the Backblaze::B2::Bucket object representing the bucket
       # specified by the required options[:b2_bucket].
       def b2_bucket
-        @b2_bucket ||= ::Backblaze::B2::Bucket.get_bucket(name: b2_credentials[:bucket])
+        @b2_bucket ||= ::Backblaze::B2::Bucket.get_bucket(
+          name: b2_credentials.fetch(:bucket)
+        )
       end
 
       # Return the specified bucket name as a String.
@@ -95,7 +92,7 @@ module Paperclip
 
       # Return whether this attachment exists in the bucket.
       def exists?(style = default_style)
-        !!get_file(filename: get_path(style))
+        !get_file(filename: get_path(style)).nil?
       end
 
       # Return a Backblaze::B2::File object representing the file named in the
@@ -124,9 +121,8 @@ module Paperclip
       # (Internal) Used by Paperclip to remove remote files from storage.
       def flush_deletes
         @queued_for_delete.each do |path|
-          if file = get_file(filename: path.sub(%r{\A/}, ''))
-            file.destroy!
-          end
+          file = get_file(filename: path.sub(%r{\A/}, ''))
+          file.destroy! unless file.nil?
         end
         @queued_for_delete = []
       end
@@ -139,7 +135,6 @@ module Paperclip
           local_file.write(body)
         end
       end
-
     end # module Backblaze
   end # module Storage
 end # module Paperclip
